feat(vscode_ide): Refactored folder structure
