+++ ide_extension/vscode/devsecops/src/extension.ts
@@ -2,164 +2,7 @@ import * as vscode from "vscode";
 import * as path from "path";
 import { iacScanRequest, imageScanRequest } from "./application/InitEngineCore";
 import { Docker, IOptions } from "docker-cli-js";
-
-class CategoryTreeItem extends vscode.TreeItem {
-  constructor(
-    public readonly label: string,
-    public readonly collapsibleState: vscode.TreeItemCollapsibleState,
-    public readonly children: vscode.TreeItem[]
-  ) {
-    super(label, collapsibleState);
-  }
-}
-class ScanResultItem extends vscode.TreeItem {
-  constructor(
-    public readonly label: string,
-    public readonly description: string,
-    public readonly timestamp: Date,
-    public readonly findings: FindingItem[]
-  ) {
-    super(label, vscode.TreeItemCollapsibleState.Collapsed);
-    this.tooltip = `${label} - ${findings.length} findings`;
-    this.description = `${findings.length} findings - ${timestamp.toLocaleString()}`;
-    this.iconPath = findings.length > 0 
-      ? new vscode.ThemeIcon("warning") 
-      : new vscode.ThemeIcon("pass");
-  }
-}
-class FindingItem extends vscode.TreeItem {
-  constructor(
-    public readonly finding: any // Using any for now, replace with proper type from your ScannerRes
-  ) {
-    super(finding.description || "Unknown Issue", vscode.TreeItemCollapsibleState.None);
-    this.description = finding.severity || "Unknown";
-    this.tooltip = `${finding.description}\nSeverity: ${finding.severity}\nResource: ${finding.resource || "Unknown"}`;
-    
-    // Set icon based on severity
-    switch (finding.severity?.toLowerCase()) {
-      case "critical":
-      case "high":
-        this.iconPath = new vscode.ThemeIcon("error");
-        break;
-      case "medium":
-        this.iconPath = new vscode.ThemeIcon("warning");
-        break;
-      case "low":
-        this.iconPath = new vscode.ThemeIcon("info");
-        break;
-      default:
-        this.iconPath = new vscode.ThemeIcon("question");
-    }
-  }
-}
-
-export class DevSecOpsTreeDataProvider
-  implements vscode.TreeDataProvider<vscode.TreeItem>
-{
-  private _onDidChangeTreeData: vscode.EventEmitter<
-    vscode.TreeItem | undefined | null | void
-  > = new vscode.EventEmitter<vscode.TreeItem | undefined | null | void>();
-  readonly onDidChangeTreeData: vscode.Event<
-    vscode.TreeItem | undefined | null | void
-  > = this._onDidChangeTreeData.event;
-  private categories: CategoryTreeItem[] = [];
-  private extensionPath: string;
-  private scanResults: ScanResultItem[] = [];
-
-  constructor(private context: vscode.ExtensionContext) {
-    this.extensionPath = context.extensionPath;
-    this.getItems();
-  }
-
-  public refresh(): void {
-    this.getItems();
-    this._onDidChangeTreeData.fire();
-  }
-
-  getTreeItem(element: vscode.TreeItem): vscode.TreeItem {
-    return element;
-  }
-
-  public addScanResult(label: string, findings: any[], sourceType: 'iac' | 'image'): void {
-    const timestamp = new Date();
-    const findingItems = findings.map(f => new FindingItem(f));
-    
-    this.scanResults.unshift(new ScanResultItem(
-      `${sourceType === 'iac' ? 'IaC' : 'Image'}: ${label}`,
-      sourceType === 'iac' ? 'Infrastructure as Code' : 'Container Image',
-      timestamp,
-      findingItems
-    ));
-    
-    // Keep only the last 5 scan results
-    if (this.scanResults.length > 5) {
-      this.scanResults = this.scanResults.slice(0, 5);
-    }
-    
-    this.refresh();
-  }
-
-  getChildren(element?: vscode.TreeItem): Thenable<vscode.TreeItem[]> {
-    if (!element) {
-      return Promise.resolve(this.categories);
-    }
-
-    if (element instanceof CategoryTreeItem) {
-      return Promise.resolve(element.children);
-    }
-    return Promise.resolve([]);
-  }
-
-  private getItems(): void {
-    const iacScanItems: vscode.TreeItem[] = [];
-    const imageScanItems: vscode.TreeItem[] = [];
-
-    const imageScanItem = new vscode.TreeItem(
-      "Image Scan",
-      vscode.TreeItemCollapsibleState.None
-    );
-    imageScanItem.command = {
-      command: "devsecops.imageScan",
-      title: "IMAGE_SCAN",
-      arguments: [imageScanItem],
-    };
-    imageScanItem.iconPath = new vscode.ThemeIcon("breakpoints-view-icon");
-    imageScanItem.tooltip = "Scan a docker image";
-    imageScanItems.push(imageScanItem);
-
-    const iacScanItem = new vscode.TreeItem(
-      "Infrastructure as Code Scan",
-      vscode.TreeItemCollapsibleState.None
-    );
-    iacScanItem.command = {
-      command: "devsecops.iacScan",
-      title: "IAC_SCAN",
-      arguments: [iacScanItem],
-    };
-    iacScanItem.iconPath = new vscode.ThemeIcon("breakpoints-view-icon");
-    iacScanItem.tooltip =
-      "Scan a folder for IaC vulnerabilities like k8s or dockerfiles";
-    iacScanItems.push(iacScanItem);
-
-    this.categories = [
-      new CategoryTreeItem(
-        "Infrastructure as code scans",
-        vscode.TreeItemCollapsibleState.Expanded,
-        iacScanItems
-      ),
-      new CategoryTreeItem(
-        "Containers scans",
-        vscode.TreeItemCollapsibleState.Collapsed,
-        imageScanItems
-      ),
-      new CategoryTreeItem(
-        "Scan results",
-        vscode.TreeItemCollapsibleState.Expanded,
-        this.scanResults
-      ),
-    ];
-  }
-}
+import { DevSecOpsTreeDataProvider } from "./tree/DevSecOpsTreeDataProvider";
 
 export function activate(context: vscode.ExtensionContext) {
   const treeDataProvider = new DevSecOpsTreeDataProvider(context);
