+++ ide_extension/vscode/devsecops/src/tree/results/ScanResultItem.ts
@@ -1,0 +1,18 @@
+import * as vscode from "vscode";
+import { FindingItem } from "./finding/FindingItem";
+
+export class ScanResultItem extends vscode.TreeItem {
+  constructor(
+    public readonly label: string,
+    public readonly description: string,
+    public readonly timestamp: Date,
+    public readonly findings: FindingItem[]
+  ) {
+    super(label, vscode.TreeItemCollapsibleState.Collapsed);
+    this.tooltip = `${label} - ${findings.length} findings`;
+    this.description = `${findings.length} findings - ${timestamp.toLocaleString()}`;
+    this.iconPath = findings.length > 0 
+      ? new vscode.ThemeIcon("warning") 
+      : new vscode.ThemeIcon("pass");
+  }
+}
